fix: intercom
